requirements.txt
Django
djangorestframework
django-cors-headers
channels
channels-redis
celery
redis
django-environ
mysqlclient
daphne